refactor: typescript fixes
