+++ frontend/src/components/Icon.tsx
@@ -5,7 +5,12 @@ const StyledIcon = styled.span`
   user-drag: none !important;
 `;
 
-const Icon = ({ icon, style }) => {
+interface IconProps {
+  icon: string;
+  style?: React.CSSProperties;
+}
+
+const Icon = ({ icon, style }:IconProps) => {
   return (
     <StyledIcon className="icon material-symbols-outlined" style={style}>
       {icon}
