+++ frontend/vite.config.ts
@@ -10,6 +10,15 @@ export default ({ mode }: ConfigEnv) => {
   const SERVER_URL = process?.env?.SERVER_URL || 'http://localhost:4455';
 
   return defineConfig({
+    resolve: {
+      alias: {
+        '@containers': '/src/containers',
+        '@components': '/src/components',
+        '@hooks': '/src/hooks',
+        '@types': '/src/types',
+      },
+    },
+
     server: {
       proxy: {
         '/api': {
